Add utility to check sdk version, to be used by mainline modules.
There are a few variations of these already, as modules start
developing features. There seems to be a consensus that it would be good
to have a common library that all modules use and that Mainline team
ensures is safe to release monthly and is properly updated when S SDK is
finalised.

Bug: 170831689
Test: Used in MediaProvider on build from sc-release.
Change-Id: Ie5c6eb44024d423fc65b7db7c6b0ed050ddcd785