Move BasicShellCommandHandler
The class is currently used by the framework and mainline modules,
including telephony, wifi mainline module and the incoming
ConnectivityService module. For these kind of class that are shared
between the framework and modules should be placed in modules-utils.
Thus, move this class to this folder which was created for putting
such utils class. Also, classes inside framework(e.g. ShellCommand)
extend BasicShellCommandHandler, and it uses these two parent's
members. But they are now in different packages after
BasicShellCommandHandler is moved to modules-utils. Thus, also
update TAG and DEBUG to protected.

Bug: 170598012
Test: m
Change-Id: Ia04c7115924179540ef227cd5f50915e206c8968
Ignore-AOSP-First: Other commit in the same topic got code conflict
between aosp and internal, so I have to merge in internal first.