Make framework-annotations visible to libcore/ and art/
libcore will annotate its APIs as @SystemApi and thus needs
access to SystemApi.java. libcore can't depend on
framework-annotations-lib because it depends on core-current SDK which
requires libcore. Use the file group to break the loop.

Bug: 157639992
Test: m droid
Ignore-AOSP-First: Submitting on AOSP causes merge conflict, because a large number of CLs are missing the downstream branch git_mainline-prod.
Change-Id: I281f0cd7a55bccc7a27681c7efb700a48e5bfcca