Move annotations in public api to frameworks/base
These two are the only annotations in the android.annotations package
that are not @hide -- they are actually in the public API.

As such, they are not suitable in the annotations library. Move them
back to frameworks/base where the other core java APIs live.

Bug: 186197911
Test: m
Merged-In: Idec38b6cb691e4bec13e7efdc915ac6ad28a531f
Change-Id: Idec38b6cb691e4bec13e7efdc915ac6ad28a531f