@UserHandleAware parameters for permissions & sdk
As we upgrade old non-@UserHandleAware methods to become
@UserHandleAware, we must document when that behaviour-changing
modification was made. We also need to document permissions requirements
that only apply if the context and calling users differ (to avoid making
it appear that new permission requirements are being demanded even in
the original calling-user case).

We accomplish this by adding some parameters to the annotation.

Test: atest com.android.server.pm.UserManagerTest
Test: atest android.multiuser.cts.UserManagerTest
Bug: 183155436
Merged-In: Id711655a3d464ccdbdd4a35425c085e594088caf
Change-Id: Id711655a3d464ccdbdd4a35425c085e594088caf