Add @EnforcePermission annotation
Introduce a new annotation targeting system services. Similarly to
@RequiresPermission, it indicates the permission(s) required by a method
when called. As opposed to @RequiresPermission, the verification is
strictly enforced on the server side, based on the exact expression of
the annotation.

This Java annotation is automatically generated for AIDL interfaces that
use the corresponding AIDL @Enforce annotation. The intent is to declare
the required permission(s) when defining an interface and consistently
implement the associated verification.

Bug: 197828948
Test: atest --host aidl_unittests
Change-Id: I7097322fcd81b1e38e8931b757030e174b749d00