Add static list to reuse existing SynchronousResultReceiver
When instantiating a new SynchronousResultReceiver,
BinderProxy keeps a reference to its instance. As Bluetooth
makes a lot of calls to its service with this class, we
need to store and reuse instead of instantiate each time.
Tested with 4 process each creating 4 threads each calling
30000 times a service method through SyncResultReceiver,
all calls with a param value that is returned and checked
once the result is retrieved.

Bug: 217366135
Test: manual
Tag: #feature
Ignore-AOSP-First: cherry-pick
Change-Id: Ide3336c624a0e58896072656a2458f25fb0a3572