Add FlaggedApiRule
Introduce a new JUnit rule, FlaggedApiRule, to allow unit tests
annotated @FlaggedApi to be skipped (if SKIP_TESTS_ANNOTATED_FLAGGED_API
is true). This is useful if the tests rely on @FlaggedApi symbols that
have been flagged off and are no longer accessible to the test.

To use, tests soong module should

  - update static_libs to include "modules-utils-testing" and "framework-annotations-lib"
  - create a class filed of type FlaggedApiRule
  - annotate select tests with @FlaggedApi
  - Set SKIP_TESTS_ANNOTATED_FLAGGED_API to true on select builds

A future CL may teach soong to modify SKIP_TESTS_ANNOTATED_FLAGGED_API
as part of the build.

Bug: 267288566
Test: atest com.android.modules.testing.utils.FlaggedApiRuleTest
Change-Id: Ic923337e626d33d889c69e1f7cb696e0e60badb1