@FlaggedApi: add "flag" argument
The old use of @FlaggedApi, for mainline moduels to hide APIs when
boarding trains, is not used. Repurpose @FlaggedApi to be used with
aconfig flags instead.

Add an argument to the annotation that says which flag this API is
guarded by. This annotation will be included in the API text tracking
files (current.txt et al).

Bug: 285075515
Test: m
Change-Id: I234824a2357201193afa4c46f989a8c0ffaa4918