Address visibility issues of StateInfo members
These two members are assigned after creation and read on
another thread without synchronization. This creates a
visibility issue where the pointer might be seen uninitalized.
Making them final guarantees the members are seen if the
pointer to the StateInfo object is seen.

active is changed dynamically, but its default value is
false : as long as it's only accessed on one thread (which
is the case), before the first write that thread may see
either the initialized value (false) or the default value
(also false), this is therefore safe.

Test: TH, there isn't much to test ; no more occurrences of the
      crash is the only proof
Bug: 282533638
Change-Id: I3e6b224812d52137ddb7b5b0be5e8aca95beaec1