aconfig-annotations-lib: do not depend on current Android APIs
aconfig-annotations-lib sets "sdk_version: system_current", which means
the library has an (implicit) dependency on
system-api-stubs-docs-non-updatable.

Soong will automatically add a dependency on aconfig-annotations-lib to
every java_aconfig_library. This means that everyh java_aconfig_library
has an (implicit) dependency on system-api-stubs-docs-non-updatable.

In order for metalava to correctly expand the aconfig auto-generated
Flags.FLAG constants when generating API signature files, it needs
access to the defintion of the Flags class(es). This is handled by
adding the relevant java_aconfig_library{.generated_srcjars} as sources.

In the case of system-api-stubs-docs-non-updatable, adding those sources
creates a dependency cycle: in short system-api-stubs-docs-non-updatable
depends on a java_aconfig_library, which in turn depends on
system-api-stubs-docs-non-updatable.

Break this dependency cycle by changing aconfig-annotations-lib to set
"sdk_versions: core_current".

Note: the new value of sdk_version is somewhat arbitrary. It can be
anything (e.g. "34") that doesn't involve the current Android SDK.

Bug: 301859633
Test: m nothing
Ignore-AOSP-First: part of topic that needs to be merged internally before CP'd to AOSP
Change-Id: I7f5249ec23802720d4ef6be865ef41266334d7df